Combine the ordering and tree referrals into a single reducer
