+++ public/components/CampaignAnalytics/Analytics/CampaignReferrals.js
@@ -2,7 +2,6 @@ import React from "react";
 import {connect} from "react-redux";
 import {bindActionCreators} from "redux";
 import {getCampaignReferrals} from "../../../actions/CampaignActions/getCampaignReferrals";
-import CampaignReferral from "./CampaignReferral";
 import CampaignReferralHeader from "./CampaignReferralHeader";
 import InfinityMenu from "react-infinity-menu";
 import ProgressSpinner from "../../utils/ProgressSpinner";
@@ -17,81 +16,17 @@ class CampaignReferrals extends React.Component {
     if (nextProps.campaign.id !== this.props.campaign.id) {
       this.props.campaignReferralActions.getCampaignReferrals(nextProps.campaign.id);
     }
-
-    this.addTreeToState(nextProps.campaignReferrals, nextProps.campaignReferralOrder);
   }
 
-  buildTree = (referrals, ordering) => {
-
-    let len = referrals.length,
-      tree = [],
-      i;
-
-    for (i = 0; i < len; i += 1) {
-
-      const children =
-        (referrals[i] && referrals[i].children && referrals[i].children.length > 0) ?
-          this.buildTree(referrals[i].children, ordering) :
-          [];
-
-      tree.push({
-        "name": referrals[i].sourceDescription,
-        "id": i,
-        "impressionCount": referrals[i].stats.impressionCount,
-        "clickCount": referrals[i].stats.clickCount,
-        "ctr": referrals[i].stats.ctr,
-        "customComponent": CampaignReferral,
-        "isOpen": false,
-        "children": children
-      });
-    }
-
-    // Sort children
-    if (ordering) {
-      const {order, field} = ordering;
-
-      tree.sort((a, b) => {
-
-        const compare = (numA, numB) => {
-          const invert = order === 'asc' ? 1 : -1;
-          if (numA < numB) {
-            return -1 * invert;
-          } else if (numA > numB) {
-            return invert;
-          } else {
-            return 0;
-          }
-        };
-
-        switch (field) {
-          case 'impressions':
-            return compare(a.impressionCount, b.impressionCount);
-          case 'clicks':
-            return compare(a.clickCount, b.clickCount);
-          case 'ctr':
-            return compare(a.ctr, b.ctr);
-        }
-      });
-    }
-
-    return tree;
-  };
-
-  addTreeToState = (referrals, ordering) => {
-    this.setState({
-      tree: this.buildTree(referrals, ordering)
-    });
-  };
-
-  onNodeMouseClick(event, tree) {
-    this.setState({
-      tree: tree
-    });
+  onNodeMouseClick(event, tree, node, level, keyPath) {
+    console.log('on node mouse click ', event, tree, node, level, keyPath);
   }
 
   render() {
 
-    if (this.state && this.state.tree && this.state.tree.length === 0) {
+    if (this.props.campaignReferrals &&
+        this.props.campaignReferrals.tree &&
+        this.props.campaignReferrals.tree.length === 0) {
       return (
         <div className="campaign-info campaign-box">
           <div className="campaign-box__header">Referrals from on-platform</div>
@@ -101,14 +36,14 @@ class CampaignReferrals extends React.Component {
       )
     }
 
-    if (this.state && this.state.tree) {
+    if (this.props.campaignReferrals && this.props.campaignReferrals.tree) {
       return (
         <div className="campaign-info campaign-box">
           <div className="campaign-box__header">Referrals from on-platform</div>
           <div className="campaign-box__body">
             <div className="campaign-referral-list campaign-assets__field__value">
               <InfinityMenu
-                tree={this.state.tree}
+                tree={this.props.campaignReferrals.tree}
                 disableDefaultHeaderContent={true}
                 headerContent={CampaignReferralHeader}
                 onNodeMouseClick={this.onNodeMouseClick.bind(this)}
@@ -132,8 +67,7 @@ class CampaignReferrals extends React.Component {
 
 function mapStateToProps(state) {
   return {
-    campaignReferrals: state.campaignReferrals,
-    campaignReferralOrder: state.campaignReferralOrder
+    campaignReferrals: state.campaignReferrals
   };
 }
